backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.5
pydantic==2.8.2
python-multipart==0.0.9
openai==1.40.3
numpy==1.26.4
faiss-cpu==1.8.0.post1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
